Add copyright, type hints, docstrings. use rpyc.exposed diecorators rather than 'exposed_' prefix which allows better type checking between proxies and services. Split SynthDriverService.registerNotification into registerSynthIndexReachedNotification and registerSynthDoneSpeakingNotification.
