+++ source/_bridge/runtimes/synthDriverHost/speechDictHandler.py
@@ -1,0 +1,7 @@
+# A part of NonVisual Desktop Access (NVDA)
+# Copyright (C) 2025 NV Access Limited.
+# This file may be used under the terms of the GNU General Public License, version 2 or later, as modified by the NVDA license.
+# For full terms and any additional permissions, see the NVDA license file: https://github.com/nvaccess/nvda/blob/master/copying.txt
+
+
+# Only required as synthDriverHandler imports this module, but is not used in this runtime.
