+++ source/_bridge/runtimes/synthDriverHost/synthSettingsRing.py
@@ -1,3 +1,8 @@
+# A part of NonVisual Desktop Access (NVDA)
+# Copyright (C) 2025 NV Access Limited.
+# This file may be used under the terms of the GNU General Public License, version 2 or later, as modified by the NVDA license.
+# For full terms and any additional permissions, see the NVDA license file: https://github.com/nvaccess/nvda/blob/master/copying.txt
 
+# Only required as synthDriverHandler imports this symbol, but is not used in this runtime. 
 SynthSettingsRing = None
